Moved Draw calls and added in mv and proj for shadows
All of the drawing that occured in render has now been moved to a new
functiondrawModels(commonMat); drawShadows has been started, but needs
to be finished out with the rest of the shadow drawings.

Orthographic projection matrix has been created as well as the modelview
matrix for shadows. All shadow code is still commented out